made all of the subclasses ToDo, Deadline and Event public
