+++ src/main/java/Deadline.java
@@ -1,0 +1,17 @@
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
+
+public class Deadline extends Task {
+    private LocalDateTime by;
+    public Deadline (String deadline, String by) {
+        super(deadline);
+        DateTimeFormatter inputFormatter = DateTimeFormatter.ofPattern("dd/MM/yyyy HHmm");
+        this.by = LocalDateTime.parse(by, inputFormatter);
+    }
+
+    @Override
+    public String toString() {
+        DateTimeFormatter outputFormatter = DateTimeFormatter.ofPattern("dd MMM yyyy hh:mm a");
+        return "D" + super.toString() + " | " + this.by.format(outputFormatter);
+    }
+}
