+++ src/main/java/Event.java
@@ -1,0 +1,19 @@
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
+
+public class Event extends Task {
+    private LocalDateTime from;
+    private LocalDateTime to;
+    public Event(String event, String from, String to) {
+        super(event);
+        DateTimeFormatter inputFormatter = DateTimeFormatter.ofPattern("dd/MM/yyyy HHmm");
+        this.from = LocalDateTime.parse(from, inputFormatter);
+        this.to = LocalDateTime.parse(to, inputFormatter);
+    }
+
+    @Override
+    public String toString() {
+        DateTimeFormatter outputFormatter = DateTimeFormatter.ofPattern("dd MMM yyyy hh:mm a");
+        return "E" + super.toString() + " | " + this.from.format(outputFormatter) + " - " + this.to.format(outputFormatter);
+    }
+}
