+++ src/main/java/ToDo.java
@@ -1,0 +1,11 @@
+public class ToDo extends Task {
+    public ToDo(String todo) {
+        super(todo);
+    }
+
+    @Override
+    public String toString() {
+        return "T" + super.toString();
+    }
+}
+
